feat: add window function based analytics

+++ src/schemas/order_payments_schema.py
@@ -5,5 +5,5 @@ order_payments_schema = StructType([
     StructField("payment_sequential", IntegerType(), True),
     StructField("payment_type", StringType(), True),
     StructField("payment_installments", IntegerType(), True),
-    StructField("payment_value", IntegerType(), True)
+    StructField("payment_value", FloatType(), True)
 ])
